fix: update default values of the chart helm
Change default value of the k8sDns field of the DNS disruption to allow (by default) all k8s DNS resolution to not kill all resolutions.
Remove the value of the dnsServer field of the DNS disruption because the k8s-DNS should take care of external DNS resolutions.

Jira: CHAOS-274