CHAOS-399: Implement a disk failure based on eBPF (#638)
feat: disk failure based on eBPF

Runs an eBPF program used to intercept system calls and inject errors. It is used to prevent certain processes from accessing certain files. It defines a target process and a filter path, and if the process is trying to open a file that matches the filter path, an -ENOENT error will be injected, preventing the process from opening the file.

Jira: CHAOS-399

---------

Co-authored-by: Maxime VISONNEAU <maxime.visonneau@datadoghq.com>
Co-authored-by: Antoine Gaillard <antoine.gaillard@datadoghq.com>
Co-authored-by: Nathan Tournant <ntournant81@gmail.com>
Co-authored-by: Nathan Tournant <nathan.tournant@datadoghq.com>