CHAOSPLT-124: fix - update disruption target status with multiple kinds (#807)
* fix: update disruption target status with multiple kinds

The target disruption status was not taking into account when a disruption injects multiple kinds of disruptions. Now the target status of the disruption lists the status of all its chaos pods.

* refacto: change targetInjectorList to Map

In order to reduce the complexity, with a map no need to parse each item list to find the type of the injector, it's possible to directly get the injector form its injection type. So the complexity goes from O(n) to O(1).

Jira: CHAOSPLT-124